#93 Fix where I fetch value from
Values for this widget must be fetched via
gtk_editable_get_text and then parsed.

This provides a default parseInput proc that users can override.

However this state still doesn't update the value displayed by the
widget.
It does however update the value within the state of the example so... yay?