Add basics to get page setup unix started
This provides the barebones bindings
needed to get this dialog going.
It also provides a minimal example and a starting point for the widget definition.

When this is done it should produce a GtkPageSetup.

This also should provide bindings for PageSetup to:
- load one from a file/keyfile
- manipulate it
- save it to file/gvariant/keyfile

Further it should provide bindings to setup
and customize `PaperSize` structs
as well as provide memory management for it.